Revisar TLB

+++ cpu/include/mmu.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 extern int tamanio_pagina;
-uint16_t pagina_direccion_logica(uint32_t direccion_logica);
+uint32_t pagina_direccion_logica(uint32_t direccion_logica);
 uint16_t desplazamiento_direccion_logica(uint32_t direccion_logica);
 uint8_t cantidad_paginas_enviar(int cantidad_bytes, uint32_t direccion_logica);
 #endif
